# Conflicts:
#	quski-oro-datamodel-ejb/src/main/java/com/relative/quski/model/TbQoCliente.java
#	quski-oro-service-ejb/src/main/java/com/relative/quski/service/QuskiOroService.java

+++ quski-oro-datamodel-ejb/src/main/java/com/relative/quski/model/TbQoCliente.java
@@ -5,7 +5,6 @@ import javax.persistence.*;
 
 import com.relative.quski.enums.EstadoEnum;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -36,7 +35,6 @@ public class TbQoCliente implements Serializable {
 	@Column(name = "apellido_paterno")
 	private String apellidoPaterno;
 
-
 	private String campania;
 
 	@Column(name = "canal_contacto")
@@ -51,7 +49,7 @@ public class TbQoCliente implements Serializable {
 	@Column(name = "cedula_cliente")
 	private String cedulaCliente;
 
-	private BigDecimal edad;
+	private Long edad;
 
 	private String email;
 	
@@ -83,14 +81,12 @@ public class TbQoCliente implements Serializable {
 	@Column(name = "nivel_educacion")
 	private String nivelEducacion;
 
-
 	@Column(name = "primer_nombre")
 	private String primerNombre;
 
 	private String profesion;
 
 	private String publicidad;
-
 
 	@Column(name = "segundo_nombre")
 	private String segundoNombre;
@@ -208,8 +204,6 @@ public class TbQoCliente implements Serializable {
 		this.usuario = usuario;
 	}
 
-
-
 	public String getCampania() {
 		return this.campania;
 	}
@@ -242,11 +236,11 @@ public class TbQoCliente implements Serializable {
 		this.cedulaCliente = cedulaCliente;
 	}
 
-	public BigDecimal getEdad() {
+	public Long getEdad() {
 		return this.edad;
 	}
 
-	public void setEdad(BigDecimal edad) {
+	public void setEdad(Long edad) {
 		this.edad = edad;
 	}
 
@@ -361,7 +355,6 @@ public class TbQoCliente implements Serializable {
 	public void setPublicidad(String publicidad) {
 		this.publicidad = publicidad;
 	}
-
 
 	public String getSegundoNombre() {
 		return this.segundoNombre;
